gating diagnostics -> rustc_session::parse
